+++ radical_translations/static/js/project.js
@@ -1,5 +1,12 @@
 /* Project specific Javascript goes here. */
 $(function () {
+  // The ChartDataLabels plugin registers itself globally, meaning that once imported,
+  // all charts will display labels. We only need it enabled on specific charts, we
+  // need to unregister it globally.
+  // https://chartjs-plugin-datalabels.netlify.app/guide/getting-started.html#registration
+  Chart.plugins.unregister(ChartDataLabels)
+
+  // override the default image path for leaflet
   L.Icon.Default.imagePath = '/static/leaflet/dist/images/'
 
   // highlight active tab and allow permalink for bibliography and record pages
@@ -71,3 +78,31 @@ const getMap = (element) => {
 const dispatchWindowResizeEvent = () => {
   window.dispatchEvent(new Event('resize'))
 }
+
+/* UI for static timeline mockup
+ * TO DELETE once implemented */
+
+$(function () {
+  // Toggle contextual information
+  $('.country').on('click', function () {
+    $('.info').toggleClass('hidden')
+
+    return false
+  })
+
+  // Toggle key
+  $('.key-button').on('click', function () {
+    $('.key-info').toggleClass('hidden')
+    $('#events-key-svg').attr('viewBox', $('#events-key-svg').attr('viewBox') 
+      === '0 0 1920 100' ? '0 0 1920 300' : '0 0 1920 100')
+
+    return false
+  })
+
+  // Close event info box
+  $('.close-event').on('click', function () {
+    console.log('Chiudi!')
+    alert('Chiudi!')
+    return false
+  })
+})